backend/requirements.txt
Django==5.1
djangorestframework==3.15
djangorestframework-simplejwt==5.3
django-cors-headers==4.4
django-environ==0.11.2
celery[redis]==5.4
django-celery-beat==2.7
yfinance==0.2.48
httpx==0.27
psycopg2-binary==2.9.10
gunicorn==23.0
whitenoise==6.7
setuptools<81